+++ src/Element/Element.php
@@ -14,12 +14,6 @@ use craft\base\NestedElementInterface;
 use craft\behaviors\CustomFieldBehavior;
 use craft\db\Connection;
 use craft\db\ExcludeDescendantIdsExpression;
-use craft\elements\actions\Delete;
-use craft\elements\actions\DeleteActionInterface;
-use craft\elements\actions\Duplicate;
-use craft\elements\actions\Edit;
-use craft\elements\actions\SetStatus;
-use craft\elements\actions\View as ViewAction;
 use craft\elements\db\NestedElementQueryInterface;
 use craft\events\AuthorizationCheckEvent;
 use craft\events\DefineAttributeKeywordsEvent;
@@ -72,7 +66,6 @@ use DateTime;
 use Deprecated;
 use GraphQL\Type\Definition\Type;
 use Illuminate\Database\Query\Builder;
-use Illuminate\Support\Collection;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Validator as ValidatorFacade;
 use Illuminate\Support\Traits\Macroable;
@@ -146,6 +139,7 @@ abstract class Element extends Component implements ElementInterface
     }
     use Concerns\Eagerloadable;
     use Concerns\Exportable;
+    use Concerns\HasActions;
     use Concerns\HasControlPanelUI;
     use Concerns\HasCustomFields;
     use Concerns\HasSources;
@@ -810,100 +804,6 @@ abstract class Element extends Component implements ElementInterface
             self::STATUS_ENABLED => t('Enabled'),
             self::STATUS_DISABLED => t('Disabled'),
         ];
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public static function actions(string $source): array
-    {
-        $actions = Collection::make(static::defineActions($source));
-
-        $hasActionType = fn (string $type) => $actions->contains(
-            fn ($action) => (
-                $action === $type ||
-                $action instanceof $type ||
-                is_subclass_of($action, $type) ||
-                (
-                    is_array($action) &&
-                    isset($action['type']) &&
-                    ($action['type'] === $type || is_subclass_of($action['type'], $type))
-                )
-            ),
-        );
-
-        // Prepend Duplicate?
-        if (! $hasActionType(Duplicate::class)) {
-            $actions->prepend(Duplicate::class);
-        }
-
-        // Prepend Edit?
-        if (! $hasActionType(Edit::class)) {
-            $actions->prepend([
-                'type' => Edit::class,
-                'label' => mb_ucfirst(t('Edit {type}', [
-                    'type' => static::lowerDisplayName(),
-                ])),
-            ]);
-        }
-
-        // Prepend View?
-        if (static::hasUris() && ! $hasActionType(ViewAction::class)) {
-            $actions->prepend([
-                'type' => ViewAction::class,
-                'label' => mb_ucfirst(t('View {type}', [
-                    'type' => static::lowerDisplayName(),
-                ])),
-            ]);
-        }
-
-        // Prepend Set Status?
-        if (static::includeSetStatusAction() && ! $hasActionType(SetStatus::class)) {
-            $actions->prepend(SetStatus::class);
-        }
-
-        // Append Delete?
-        if (! $hasActionType(DeleteActionInterface::class)) {
-            $actions->push(Delete::class);
-        }
-
-        $actions = $actions->all();
-
-        // Fire a 'registerActions' event
-        if (Event::hasHandlers(static::class, self::EVENT_REGISTER_ACTIONS)) {
-            $event = new RegisterElementActionsEvent([
-                'source' => $source,
-                'actions' => $actions,
-            ]);
-            Event::trigger(static::class, self::EVENT_REGISTER_ACTIONS, $event);
-
-            return $event->actions;
-        }
-
-        return $actions;
-    }
-
-    /**
-     * Returns whether the Set Status action should be included in [[actions()]] automatically.
-     *
-     * @since 4.3.2
-     */
-    protected static function includeSetStatusAction(): bool
-    {
-        return false;
-    }
-
-    /**
-     * Defines the available bulk element actions for a given source.
-     *
-     * @param  string  $source  The selected source’s key, if any.
-     * @return array The available bulk element actions.
-     *
-     * @see actions()
-     */
-    protected static function defineActions(string $source): array
-    {
-        return [];
     }
 
     /**
